requirements.txt
asgiref==3.4.1
certifi==2021.5.30
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
gunicorn==20.1.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
pandas==1.3.4
python-binance==1.0.15
tweepy==4.4.0
